Move gloas http logic to modules.
