+++ beacon_node/http_api/src/beacon/mod.rs
@@ -1,2 +1,3 @@
+pub mod execution_payload_envelope;
 pub mod pool;
 pub mod states;
